requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-mail==0.10.0
flask-bcrypt==1.0.1
flask-wtf==1.2.1
itsdangerous==2.2.0
tensorflow==2.17.0
numpy==1.26.4
pillow==10.4.0
python-dotenv==1.0.1
reportlab==4.2.2
wtforms==3.1.2
email-validator==2.2.0